Add script to create forecast.ss file
Script will generate a forecast file that can be used within ss3sim.
By default forecasts are turned off (as they have to be in an OM) but
can easily be turned on through arguments if you want a forecast file
that produces forecasts. This will help standardize forecast files
across ss3models.