Update casefiles to use system.file to find cases dir
For developers, you must have loaded the package using load_all
otherwise the cases will be placed in the library where the package
is loaded. I added an if statement for my personal computer but I am
not sure how to get R to recognize where the package may have been
installed from.